+++ frontend/src/app/engagement_request/page.tsx
@@ -1,5 +1,5 @@
 import DashboardLayout from "@/shared/components/DashboardLayout";
-import EngagementHome from "./engagement_home";
+import EngagementHome from "./engagement_request";
 
 
 export default function EngagementsPage() {
